+++ CHANGES.md
@@ -5,6 +5,7 @@
 ##### Fixes :wrench:
 
 - Fixed a bug introduced in v0.7.0 where Bing credits were not being collected.
+- Fixed a bug where disabling frustum culling caused external tilesets to not load.
 
 ### v0.7.0 - 2021-09-01